move Fixnum & RblFloat from prim/* to root of src
