+++ roscala/src/main/scala/coop/rchain/rosette/Meta.scala
@@ -1,7 +1,5 @@
 package coop.rchain.rosette
 
-import coop.rchain.rosette.prim.RblFloat
-
 object Meta {
   case class StdMeta(override val extension: StdExtension = null)
       extends Actor {
