+++ roscala/src/test/scala/coop/rchain/rosette/PrimPropertySpec.scala
@@ -1,6 +1,5 @@
 package coop.rchain.rosette
 
-import coop.rchain.rosette.prim.Fixnum
 import org.scalacheck.Gen
 import org.scalatest._
 import org.scalatest.prop.PropertyChecks
