+++ roscala/src/test/scala/coop/rchain/rosette/TransitionSpec.scala
@@ -2,7 +2,6 @@ package coop.rchain.rosette
 
 import coop.rchain.rosette.Meta.StdMeta
 import coop.rchain.rosette.expr.{LetExpr, RequestExpr, TupleExpr}
-import coop.rchain.rosette.prim.{Fixnum, RblFloat}
 import org.scalatest._
 
 /**
